PR-SUG: introduce custom health serializer
